chore(gatsby) Rewrite requires-writer.js to TypeScript (#24289)
- Mostly just adds TypeScript type annotations

- Introduces one assertion for matchPath !== undefined
as IGatsbyPage has it as string | undefined, but everything
in requires-writer.js assumes that it's a string. It would throw an
error anyway, I just made a bit more pretty with reporter.panic()

WARNING:
- requires-writer uses reporter from gatsby-cli, and not the
TS source, but the compiled version. Not sure if it's ok.
Maybe we should extract reporter into a separate package if
it is used in both gatsby and gasby-cli